Modify fuse_inout test not use be::engine::Engine
In the tests, the `fe::fuse::Fs` object is created with an actual
store and catalog, not with handles to a `be::engine::Engine`. This
allows testing the `Fs` object independenly of the `Engine` and also
allows testing in in memory catalogs and
stores (`be::catalog::MemCatalog` and `be::store::MemStore`), which
are not compatible with the `Engine`.